+++ pagegraph/__init__.py
@@ -1,4 +1,4 @@
 from packaging.version import Version
 
 __author__ = "Peter Snyder (pes@peteresnyder.com)"
-__version__ = Version("0.9.5")
+__version__ = Version("0.9.6")